Add bindable property to wrap a control's value (#119)
* Provide a `setUp` to properly configure the associated property

The introduced `setUp` is extremely helpful to setup any signals that may
affect the property once.

* Add bindable property to wrap a control's value

This property is common to every `UIControl` and can be used by
providing a getter and a setter for the value wrapped. This property
uses `UIControlEvents.ValueChanged` and `UIControlEvents.EditingChanged`
events to detect changes and keep the value up-to-date. This kind of
logic can be reused instead of being defined and used for each control.

* Use `setUp` to install the dependent signal of `rex_dismissAnimated`

This prevents repetition of events in cases where the property is
accessed more than once (each access adds a new `rac_signalForSelector`
for the dismiss of the view controller).

* Fixes #129 - UITextField's text bindable property should be of optional type