Allow for observable enum values (#3667)
* Fix method name in describe block

* Allow for observable enum values

Previously, when using the `NSObject.producer/signal(for:)` syntax with Swift
4 KeyPaths, we were force-casting the resulting value to the expected type.
This works for native Objective-C types, but breaks for Objective-C enums,
which are sent as their raw Integer types. This means the force-cast would
result in a crash if you tried to observe an enum value.

To fix this, we can add a special override for the KeyPath observation API
that handles the RawRepresentable conversion automatically. This will then be
used if the user specifies a KeyPath that points to a RawRepresentable type.